Move ball class to own module
